// frontend/gear-vault-web/src/api/generated/client/models/UpdateVendorDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateVendorDto = {
    name?: string | null;
    contactEmail?: string | null;
    phone?: string | null;
    address?: string | null;
    isActive?: boolean | null;
};